Implemented python RunSolver script (#199)
* Implemented python RunSolver script

* Changelog

* Fixed some bugs

* Fixing various bugs

* Updating solver wrappers to allow for log streams

* Updating solver wrappers example

* Updating pyrunsolver to also capture output streams correctly from child and pass to file/terminal

* bugfixes and test updates for pyrunsolver

* remove old comment

* Bug fixes for pytest

---------

+++ sparkle/CLI/add_solver.py
@@ -129,9 +129,13 @@ def main(argv: list[str]) -> None:
             f"{solver_source.name}. This will be replaced with "
             f"Sparkle's version of RunSolver. ({runsolver_path})"
         )
-    runsolver_target = solver_directory / runsolver_path.name
-    shutil.copyfile(runsolver_path, runsolver_target)
-    runsolver_target.chmod(os.stat(runsolver_target).st_mode | stat.S_IEXEC)
+
+    if runsolver_path.exists():
+        runsolver_target = solver_directory / runsolver_path.name
+        shutil.copyfile(runsolver_path, runsolver_target)
+        runsolver_target.chmod(os.stat(runsolver_target).st_mode | stat.S_IEXEC)
+    else:
+        print("Warning! RunSolver does not exists. Falling back to PyRunSolver.")
 
     performance_data = PerformanceDataFrame(
         gv.settings().DEFAULT_performance_data_path,
